+++ src/dbt_score/cli.py
@@ -7,8 +7,8 @@ import click
 from click.core import ParameterSource
 from dbt.cli.options import MultiOption
 
-from dbt_score.dbt_utils import dbt_parse, get_default_manifest_path
 from dbt_score.lint import lint_dbt_project
+from dbt_score.parse import dbt_parse, get_default_manifest_path
 
 BANNER: Final[str] = r"""
           __ __     __
